refactor project structure to be correctly 'go-like'
