+++ assets/goose_header.txt
