+++ internal/pkg/entropy/entropy.go
